.vimrc: Copied from cjs pre-dh configuration
This is far too large and complex for cjs-minimal; we're going to need
to strip it down and break out a lot of pieces into separate
configuration for cjs-standard or whatever we'll call the full
non-desktop configuration.

However, for the moment it will work for sharing with folks who have
no real vim configuration of their own.