.vimrc: Don't set (override) $BASH_ENV
This should be set appropriately in the parent environment since
whatever set that up knows whether, e.g., `.bashrc` knows to DTRT
when called in a non-interactive environment.

Just now it was breaking things because the parent environment started
using a different value for `$BASH_ENV` and this overrode that.