dh/update: Remove
That was an attempt to let the local repo keep up to date with changes
to both the "real origin" remote and the alternate remote we use when
we can't push to the real origin. However, we were doing this wrong.

What we want to do is make the upstream of our master branch the
origin repo master branch, whether or not we can push to it.

When we can't push changes to the real origin, we push them up to the
alternate repo with an explicit `push alternate-remote master`,
leaving both us and the alternate repo with the changes. The origin
repo will catch up later when it merges those commits. However, even
when the origin repo doesn't have them, `pull --rebase` from the
origin keeps our local changes.

The main downside of this is that `git status` shows us as being ahead
of our upstream even after we've pushed up the changes to the
alternate repo. Possibly our custom `st` git status command can be
updated to give an indication when there are changes not pushed up to
the alternate repo, as well as the expected changes not on the real
origin master branch.

When another repo wants the changes on the alternate before they get
to the origin, it fetches and then rebases on to the alternate master.
(This is also how we get changes to push up to the real origin.)

This relies on the alternate repo's master branch being quickly
updated when new commits come on to the real origin's master branch,
but this can easily be done by anybody who can push to the alterate
origin.