.vimrc: set shellcmdflag=-i\ -c so shells are "interactive"
This gives us all our functions, aliases, etc. when we spawn a shell
with `:!` or other commands.

This was previously done by setting `BASH_ENV=~/.bashrc` but that's
horribly broken because it also gets passed on to subshells of those
spawned shells which breaks things when `.bashrc` itself spawns any
subshell (such as by running a shell script).