.bashrc.inb3: prepath $HOME/.local/bin
This is necessary for xtitle to be found on systems that don't
automatically include ~/.local/bin in the path.

This is also probably the proper place to do this to make sure that
it's in the path at all, though it's likely that less minimal startup
scripts would want to do this again near the end to force it to the
front.