.vimrc: unmap `gn` and `gN`
`gn`, a long time ago, turned off search highlighting and has since
been replaced by `gu`. `gN` moved to the next error which I've not
used in years.

Removing these unmasks the default functions which are to search
forward or backward for the next search pattern and then make it the
visual mode selection.