.gitconfig: Set push.recurseSubmodules = check
We can't think of any particular reason not to go with `on-demand`
since if you got as far as committing in the sub-repo you know you're
going to have to push and if you just `push` after a check it does
just what on-demand would have done anyway. But for some reason we
still don't feel entirely comfortable with it, so we simply halt and
force the user to do it manually (unless he's overridden it in later
config).