.vimrc: map gp gqap
This does not override any existing vim commands.

Saves a keystroke on my most common reformat operation, reformat
current paragraph, bringing it down to the same number of keystrokes
as the old version, `*f` (which included a shift keystroke).

You'd think that one more keystroke on top of three already wouldn't
make that much difference, but it seems to.

I considered `gP` but it's slightly more awkward (due to the shift)
and also overrides an existing vim paste command that I don't use now
but could see myself learning and using in the future.