.vimrc: Change showbreak to ┊ and its (and all nontext) color to darkgray
The `┊` char is Unicode U+250A Box Drawings Light Quadruple Dash
Vertical. I think this should be pretty common, and it works in my
current fonts for uxvt and main xterm (misc-fixed), though not in
whatever default xterm fonts are used for the other sizes.

I also make it a bit lighter yet by changing the color to darkgray.
This actually affects all the nontext stuff (tabs, end of line, etc.)
shown in 'list' mode, which is fine.