bin/xtitle: Use `|` in title lines and add leading space when no title
Using a vertical bar to separate the sections (title, hostname, etc.)
makes the line easier to read.

Adding leading spaces when no title has been set makes for a nice
offset in a list of window titles (such as at the FVWM window list)
making it easier to quickly latch on to windows that do have assigned
titles, which are presumably the more important ones.