.vimrc: qq → regular :quit
It used to do (via the QuitSmall() function) a :quitall if we weren't
in a temporary buffer, but I keep feeling I want to `qq` to just close
any window, not close some but quit on others. So we go back to the
normal :quit command for that.

We still want a :quitall binding, though, so that we can force an exit
even if we have unseen files in the args list or multiple windows
open, so move that to `qA`. Also, remove `q<C-Q>` because we can't use
Ctrl-S or Ctrl-Q in many terminals due to them being flow control
chars.

`qw` is no longer necessary for :quit, so repurpose it to :close, as
in "quit window," though I'm not sure we'll ever use this now since I
all it does is refuse to quit if you're in the last window.