.vimrc: Remove scrolloffset, add g{ey} scroll keybindings
`g{ey}` do 6 lines of Ctrl-{EY}, which is a nice way of exposing a
missing bit of of a function at the top or bottom of a screen. Because
we can now do this so easily, we don't need to bother with having a
scrolloffset any more that was mainly there to give us a bit of
context when a search jumps.

This also removes the annoying scroll down of the original window when
the cursor is at the top scrolloffset and and I do a vertical split.
(That appears to be a bug in vim.)

Ideally, given a count, we should do count*6 lines, but I can't be
arsed to look up how to do that right now because the count at the
moment gives us 10*count + 6 which is close enough, or maybe even
better.