.vimrc: Set `comments` for Markdown formatting
When using paragraph reformatting in markdown we set `*` (bullet) to
be a first-line-only comment character so it doesn't get duplicated on
subsequent wrapped lines of the same bullet point.

We have a new mapping, `gtm` to set this, but unfortunately we
duplicate the code in the `autocmd FileType markdown`. We need to
consider better ways to handle this.