.vimrc: Add `g/` mapping to search for visually selected text
This mapping works only while the visual selection is active (`vmap`).

We yank the current selection, start a search, use Ctrl-R to insert
the contents of the `"` register (last yank) into the search line, and
then CR to actually do the search. This comes from the Vim Tips
Wiki[1], though the same macro is also shown (in a slightly different
context) in the `help visual-operators` documentation

[1]: http://vim.wikia.com/wiki/Search_for_visually_selected_text