.vimrc: Change highlighting of Markdown ID declarations
In "word processor" mode we must wrap lines, which causes the ref
declarations and their URLs to be wrapped as well. If the URLs are
long, this can make the ref IDs hard to read, so change their colour
so that they stand out amongst all the URL text.

Also document an issue with the highlighting that we need to look at
at some point.