bin/xtitle: Show ckssh compartment rather than key info
Eventually we want to indicate if the compartment has the right keys
in it or not, ckssh has to be extended to return this info first. In
the meantime we just show a count of the number of keys in the agent.

The grep for 'The agent has no identities.' is kinda hokey, but easier
and faster than actually trying to use the exit code.