.vimrc: Restore q: q/ q?, add q; to go to ex mode
`q:`, `q/` and `q?` are the only three built-in commands starting with
`q` (anything else starts recording), and these were rendered
inaccessible when I mapped `q` followed by anything not otherwise
mapped not to do anything (so as to turn off recording). So I
manually restore these.

Also, I add `q;` to go to ex mode because I'm intending to re-use the
standard version of that, `gQ`, as a prefix for a further set of
commands. (And, since it is a sort-windowing operation, in a way,
it's comfortable under the `q` prefix.)