.vimrc digraphs: Middle dot is also known as 'interpunct'
I was originally using interpuct instead of bullet in lists because it
looked better in my terminal fonts, but looking at the semantics it
appears that this might not be a good idea. So perhaps I should consider
putting bullet back on `.m` and putting interpunct on `.M` or wherever.