bin/show-terminal-colors: Demo ANSI terminal attributes and colors
This is handy when tweaking color settings for terminal emulators,
deciding what colors to use for syntax highlighting, and the like.

It's here rather than in `cjsdesktop` because it's used on remote
systems too, especially when using things like GNU Screen or tmate.

This was moved from the now-abandoned `desktop-utils` repo.