bin/distro: Show what Linux distro we're on and do generic package handling
This works on some systems, but was never completed or even well tested.

In particular, the 'install packages' part of it is such a bad hack
that we've not even brought in the configuration files that it reads.
We'll rewrite that part of it before using it again. Possibly even
that part of it should be moved to a different installation program,
such as the ones we use for nodejs, fonts, etc.