.vimrc: Enable minimal list mode by default
This just shows tabs, non-breaking spaces and EOL spaces. I
switched to a less obtrusive initial character and removed the
fill for tabs so it should be more informative than annoying (not
that I use tabs much these days anyway).

The EOL whitespace thing might be a bit more annoying; we'll see.