.vim/cjs/main: Rework `g` commands to format to beginning/end of para
The old format-to-end, `g]`, overrode the key binding to select tag
from word under cursor. Oops!

We put them both on the same key now, `[{`, and change the unshifted
version to format forward (it was previously backward), and the
shifted version now formats backward. The format-forward is slightly
less used, but it's better on a shift key because unshifted-forward
vs. shifted-backwards is a standard thing, a là forward/backward
search next with `nN`,