bin/xtitle: Add `-n` argument to ckset
The latest commit on the dev/cjs/180503/rewrite branch changes the
behaviour to add keys when no argument is given, and this caused a key
add with every prompt because ckset doesn't yet detect that the key is
already there.

This commit adds the `-n` option to suppress that behaviour. It does
break the display if using an older version of ckssh because an error
is returned instead of the compartment name, but that's ok because we
throw that away, leaving us just displaying `?` for an unknown
compartment if we happen to be using an old version.

But a standard update will update both this and ckssh to the latest
version so there will be no issue anyway.