.vim/cjs.d/main: Don't set shell bash-rcc if we're using Windows
This is a quick hack with complex implications that need to be sorted
out properly one day. It looks, for example, as if vim run in a MinGW
environment is not considered to `has('win32')` and so we need to
consider how to deal with that. (It's currently broken with the same
complaint that gvim had: `Cannot execute shell bash-rcc`.)`