share/bash-completion/xtitle: Add `xtr` completion
This one is trivial since all it needs is _command() (complete as if
starting a new command after the last option starting with `-`) from
the standard framework (/usr/share/bash-completion/bash_completion).

I'll get around to `xtitle` and `xt` if they look like they'd be used
more than once in a blue moon.