.vim/cjs/main: New status line, large rework to window split behaviour
The new status line is mainly to include an indicator of the winfixheight
setting, but we add filetype and code of cursor under char as well since
these are handy and easy to do.

We now set equalalways by default so that in the very typical case of
bringing up a list of files and using `q<C-N>` 2-3 times to give 3-4
windows, they're all reasonably sized without further commands.

To keep this behaviour from being annoying for certain windows that we
specifically made a certain size, `qf` will now fix the size of that
window so that splits/closes will never resize it. This is indicated
with "FixSz" at the end of the status line. (Yes, it's on the left
rather than the right of the keyboard, but we're running out of space
on the right.)

We maintain `qM` to "maximize all," i.e. reset sizes per equalalways,
but now make sure to restore the previous state of equalalways.

`qF` can flip equalalways in case we don't want to use it for a while.