.vim/cjs/main: Improve statusline column/row display
Move the column (and virtual column, if over tabs) display to the
left, in a fixed position right next to the character code. The rows
stay at the far right, next to the percentage position in file.
Shorten the window fixed size indicator to just "FIX", all upper-case
making it more visible.

It was a bit tricky getting the column indication to be fixed width
because formatting applied to %V appears to take effect only when it's
displayed. The solution is to surround it in a `%(…%)` group and
format that.