.vim/after/syntax/markdown: Add mapping to disable code blocks
This works around the bug in the standard syntax/markdown.vim as
described in the comments.

It's probably a bad idea to have a `gh` mapping in a completely
separate file from my other mappings 'cause it makes it hard to see
the whole list. Perhaps the code should be extracted to a function
and the mapping to call that function moved up to the main file.