.vim/cjs.d/main: Add q^J to split and move to next file, like q^N
This makes q-prefixed j and n more orthogonal. It's not clear to me if
we should actually bother keeping n/p since they are very similar to
j/k. But they are not the same: n/p move to the next/previous window,
even through vertical splits or horizontal/vertical mixes.

Given that I almost never use vertical splits and, even if I did use a
mix, it's not clear to me that I'd want to use n/p instead of h/j,
there might be better uses for those keys.