.vim/cjs.d/main: Fix way of setting guifont that could produce errors
Previously, we always set the guifont to Noto Mono 8 and then changed
it if the host is a win32 system. However, IIRc this could produce an
error message if that font didn't exist on the Windows system. Instead
set a font path only once, and make sure it includes a font always
present on Windows.

I also confirmed that `has("win32")` returns 1 even when using a
"win64" vim.