.vim/cjs.d/main: :scriptencoding utf-8
Switch the local to UTF-8 while executing this file because I use
UTF-8 chars later on. This avoids error messages in environments where
the locale is not already a UTF-8 locale, such as Windows and Linux
systems that have not executed my standard .bash_profile.

It does not solve the issue that the status line may not be displayed
correctly in non-UTF-8 environments.