.vim/cjs.d/main: gtX commands for textwidth now :setlocal
When setting the textwidth and virtualedit, typically we want it
to be changed only in the current buffer because we're doing
something special there. Thus, change `:set` to `:setlocal`.

We do not do the same for the parallel mappings for tabstop and
shifwidth because normally we change these to match another
project's coding convention and want to use that for the
subsequent files in the argslist, too.