bin/distro: Use /etc/os-release if available
This looks like it's supposed to be able to be sourced, so we source
it if it exists.

For consistency, we change our internal variable names to match what
it sets and now expect the distro identifier ($ID, which we used to
call distro name) to be lower-case. lsb_release returns a distro
identifier with an inital capital, so we just run the whole output
through `tr` to make it all lower case, on the assumption that
mid-caps, if they ever exist in lsb_release distro identifiers, would
also become lower-case as well.

This now works on some newer systems where lsb_release is not
installed by default, and should work better in general.