.vim/cjs.d/main: Ensure 'clipboard' is set to nothing
MSYS or perhaps Git for Windows seems to supply an /etc/vimrc with
some "decent settings for programming" that is used by the
command-line VIM. This does `set clipboard=unnamed`, which makes all
yank operations to the unnamed register yank to the system clipboard
register, `"*`. This is annoying since I frequently rely on it to stay
unmodified while I do some preliminary editing before pasting it, so
explicitly set 'clipboard' back to the default value of empty, where
the unnamed register is completely separate.