.vim/cjs.d/main: qvd diff etc. now start w/cursor at top of new buffer
The read command we use for these leaves the cursor at the end of what
was just read, so we explicitly move to the first line of the buffer
after reading. (`:normal gg` avoids the use of any of our aliases.)

Also ensure that list mode is off when looking at log output, since
that output includes EOL spaces that would otherwise be highlighted.