.bashrc.inb3: cjsvim() to source ~/.vim/cjs, which now sets environment
We modify .vim/cjs to be a polyglot script: it now not only documents
the shell commands to set up an environment where the cjs-configured vim
will be the default, but also can be sourced in the shell to execute those
commands. (These are all still comments to Vim.)

The new cjsvim() function just sources that file.