bin/distro: Error msg for bad arg, -h no longer returns failure
Help now prints to stdout again (this time intentionally) since it's
not an error message when printed with -h, and that now also exits
with success, rather than failure, because it's doing exactly what was
requested.

Giving a bad command line argument prints a message about this to
stderr, prints the help to stdout and returns failure. Arguably we
shouldn't be sending anything to stdout when an error occurs with the
command line arguments, though.