share/distro/packages: python3-distutils
This is required for pactivate. It's part of the standard Python
distribution, but Debian separates it out into a separate package
for whatever reason.

I just noticed that all the dev packages are tagged with Debian.* only,
not Ubuntu, but I'm leaving this for the moment while I think about this
whole thing a bit further.