.vim/cjs.d/main: Column display now put screen col before byte offset
It's now displayed as ###c#### where the first number is the column
on screen (more usually what I want to see because it's mostly used
when lining up tables) and the second is the byte offset into the line
in the current encoding.

This isn't ideal because, unlike previously, it always displays the
byte offset even when it's the same as the column number.