.vim/: Move autocmds before 'syntax on' to get hilighting working at start
When entering vim syntax highlighting was not enabled, even when adding a
`:syntax enable` or `:syntax on` before setting the colorschme as solarized
advises in its example setup calls.

The issue turned out to be the subsequent `:autocmd!` that removed the
autocmds set up by the filetype and syntax highlighting system; that
obviously needs to be much closer to the beginning of the file. So move it
up as early as possible, with only a few very basic init settings before
it. Importantly, `filetype on` still preceeds it because we use the
filetype system in some of our autocmds.

This is still all a bit shakey, but getting better (I hope).