.vim/cjs.d/main: Disable `qq` command
`q` is a prefix for a lot of things, and so it can be easy to accidentally
type two `q`s when one wants one. (I have recently been closing windows by
accident.) Thus, disable this sequence completely (producing an error
message); we now always use `qw` to close a window.

There's no replacement for the "quit window" functionality because it's
not really necessary; I have plenty of other ways to exit completely.