.vim/cjs.d/main GitCommit(): mapped to `qvc` and `qvC`
This runs a `commit` script if it can find one (to allow for custom
ways of doing commits) or, failing that, just `git commit -v -a`. The
lower-case option does `commit -f` to skip editing the commit message;
upper case will drop the `-f` option which should bring up the message
for editing.

This replaces the previous `qvc` mapping that duplicated `qv/`.