bin/0cjs.bash: Source gitcmd-abbrev.bash using full path
If ~/.home/gitcmd-abbrev/ has not been installed with dot-home, which links
~/.local/bin/gitcmd-abbrev.bash, this will fail because the dot/bashrc.inb4
doesn't specify a path. But all it does is source that one file anyway, so
we can just skip that and source it ourselves in the repo location, so that
this works even if not installed.

This could potentially fall behind if dot/bashrc.inb4 starts doing more,
but it shouldn't do that because everything is supposed to be in
gitcmd-abbrev.bash.