bin/xtitle: Commands in $() were causing entire script to fail!
I had thought that commands in subshell evals `$( ... )` did not have their
return values checked, but apparently they do. So we need to add `|| true`
to a couple of those so that the script will complete now that we are using
`-Eeuo pipefail`.

Possibly in scripts like this we should be doing a `trap ... ERR` to
alert us when something is failing.