.vim/cjs.d/main: Fix ^O behaviour after g ^R
This used to add an additional position at the start of the line to
the jumplist; that is added both by the `"+y%` yank _and_ by the `%W`
move, though the second won't add to the jumplist if the first already
added it.

Use `keepjumps` on both of these so that a single ^O will now take me
back to where I had original done the `g^R` command.

Also add a `0` move to the beginning of the line once we've copied the
URL, because that provides an indication that we copied both the
reference and the URL.