bin/watchrun: Improve separator usage
We now print just a blank line at the start of a run and the separator
(preceeded by a blank line) at the end of the run. This gets rid of the
unnecessary separator at the start but, more importantly, the separator at
the end lets us know that the tests completed and exited, rather than us
having to remember how the tests indicate that (if they even do).