dot/bashrc.*: Add __0iscommand() for zsh-compatible quiet command check
This checks to see if the given name is an alias, function or executable
command file in the path, but prints nothing to stdout, returning only
an exit code. This is compatible with Bourne shell, Bash and Zsh.
(It will print to stdout if bad options are given or whatever.)

Change `prepath $HOME/.local/bin` to use this to check for `prepath` first
so we don't get an extra "pathpath: not found" message if prepath doesn't
exist, but just our own message noting it's not been added.

Also replace a use of `type` with this. And in that one, since it's
defining the `lf` alias, make it use `command ls` as per the previous
commit.