inputrc: Make this available (but not used by default) in cjs0
This is moved from cjs1:dot/inputrc, which will be changed just to
`$include` this file. This makes this file available in guest configs
where bin/0cjs.bash can use it (with `bind -f`) as well.

This makes both my vi extensions, such as ^A/^E for emacs-mode users,
and all my other customisations other things fully available when I'm
a guest on someone else's account.