.vim//main: Change sync-all/reload command; add split and almost-all-close
The :checktime command that checks all files for modifications done
outside of Vim is moved from `qs` to Ctrl-S, which has no default binding.

This frees `qs` to be used for :split, and we add qS to do a "full
unsplit", i.e., close all windows but the current one.