.vim/colors/cjs: Search → reverse Black on White
The reverse Cyan on DarkBlue was just too shouty for me.

For some reason this produces black on a moderately dark
grey background, not on white. Dunno what's up with that,
but that's the sort of thing I'm looking for (though I'd
like a slightly lighter grey).

This is very clear on xterm; not quite as much on
*-256color terminals.