bin/xititle: $XTITLE → $LC_NOTLC_XTITLE for ssh connection passthrough
The standard sshd configuration in Debian/Ubuntu allows `LC_*`
environment variables to be passed through and set on the client, but
no others. This renaming is a hack to let our `xtitle` window name
setting be passed through so it that our window title stays the same
after logging in to a remote host.