.bashrc become(): Become another user, overridden with my environment
become() takes a username and sudos to that user, but executes bash
with `--rcfile ~cjs/.home/cjs0/full-env.bash`, which sets up my full
environment from my home dir.

This is very different from sourcing 0cjs.bash: as well as giving the
full cjs environment, it uses files from ~cjs/. The good side is that
this requires no dot-home stuff at all in the target user's home dir.
The downside is that it may set up things that the target user can't
read from ~cjs/, which isn't a problem for the .home/cjs0/ stuff
designed to be read by everyone, but not even .home/cjs1/, much less
other files, are not set up that way.

We make some attempt to mitigate this by unsetting a few env vars that
end up pointing to ~cjs/ files and directories, but no dobut other
issues will come up. We need to think on this a bit.