vim/cjs: Removed as obsolete
It feels as if it might be nice to be able to `:runtime cjs` to
convert a running Vim configuraton to mine, but that brings in all
sorts of compatibility and conflict issues that aren't understood,
so probably better just to exit and restart with `.home/cjs0/bin/vi`
instead.

The clever trick to allow `source`ing of this file is so cute that
I'm reluctant to abandon it, but my non-clever code for setting up
the environment is much more reliable.