vim/main: Better document `q` setup and add recording aliases
`q;` to play back and `q'` to record are more easily under the
fingers than `@` and `Q@`, though those mappings still remain.

`q"` to edit a register is not yet written.

It's my long-term project to learn to use recording more effectively
for some of the things where I find myself repeating a sequence of
commands.