BROKEN vim/colors/cjs: Start work on 256 color nearest matching for ctermfg etc.
`ctermfg=`, `ctermbg=` and so on take a color index from 0 up to &t_Co.
But we want to use `#rrggbb` tuples for our color specifications, so we
need a function than can take such a tuple and determine which entry in
the terminal's (typically 256) color palette is the closest match.

XXX This is the start of the work on doing that.