bin/3pinst ghcup: Overwrite existing ghc stack hook; better output format
There's no real reason we wouldn't want to update the hook to the
latest version; if we happen to be using a hook with additional or
different functionality, that should be kept in dot-home and then
compared to the updated hook.

(Yes, this kinda screws people not using dot-home, but they are
getting this script from a dot-home module. :-))

Also do all our local (to 3pinst) updates at the very end, and print a
line before them that makes it clear that these are extra changes so
that if any errors or warnings arise, it's clear they're from 3pinst
and not the ghcup installer.

We also use --no-progress-meter on the curl for the hook because the
extra half dozen lines of default output made it significantly harder
to notice the warning message if it's being overwritten.