bin/vi: When my config not found, Use `command vim`
When we've found our config we use `exec vim`, which will always search
the path, never using an alias or function. However, if we don't find
the config we try to run plain `vim` which will use an alias if present.
Change this to `command vim` so we always run the program, not any
alias.

(It's not clear how necessary this really is, actually.)