vim/main: Use system `packpath` instead of entirely clearing it
We do need to remove the user packpath entries (~/.vim/…).

The removal code requires a local (actually script-scoped, since we're not
in a function) variable. If I just substitute the filter expression as the
first arg of join(), Vim gives an error "Invalid arguments for function
join(". I don't know why.