vim/colors/{cjs,rgb256}: 256-color configuration
rgb256.vim is a set of functions to find the closest matching color in
the standard xterm-256color palette to a given #rrggbb tuple. This
turns out not be as useful as I thought, as it's easier just to select
colours directly from a palette image (references are given in the
comments). So we just do that, but preserve the conversion routines
and the setup/configuraton for them in case we want to try using this
again later.

The configurations here now are:
- 256 color, for just a editor-related things (search/column highlights,
  menus), updated for nice colours in the 256-color palette.
- 8/16 color, with the only difference being that we use DarkGray
  rather rather than Gray for comments in the latter.
- Common, for most of the syntax highlighting. Much of this is simply
  changing colors to black; the remainder is dark blue and other
  colors that are from the 8-color palette.