vim/{autocmd,haskell.vim}: Remove haskell.vim; it was broken
The special haskell.vim used here was essentially just the
`indent/haskell.vim` file from vim.org/scripts. It had various bugs
including trying to do code indentation in block comments which made
comment formatting quite painful.

This means we now have to do all indentation by hand (using spaces
where ^T doesn't work, e.g., for the 'then' on the line after an
'if'), but this is better than battling the comments problem.

We now use just the standard Haskell support that comes with Vim
with a couple of minor tweaks in the proper place for them,
after/syntax/haskell.vim. We can put further tweaks in there if
we want to bring back anything that was removed.