.vim/colors/cjs: In 256 color mode, grey status line BG for non-focused windows
The status line remains light on darkest black for the window that has
the focus. In both cases now we're not actually using the foreground
colour but fixing it to 230 (which is slightly darker than my default
terminal foreground color) so that it Just Works in situations where
I'm forcing the colors of Normal mode to that, rather than using the
default.

This uses normal mode (instead of the default reverse mode) for this;
this seems to work slightly better by avoiding interference between
colour selection and reverse mode. (The standard config probably uses
reverse mode because the colors fall back to the Normal mode colors.)