inputrc: Fix sensitive repos path
~/co/sensitive/ does not use gh/ or gl/ subdirs but instead the subdir
is name for the company or similar, rather than a public repo source.
So we need just one entry for this, and we change the key to `s` for
"sensitive." (This requires changing SSH_AUTH_SOCK to `S`.)

Since we're now doing upper case for environment variables, change
LD_LIBRARY_PATH from `p` to `P` to match. Though it feels like we want to
use easier-to-access Ctrl-P and Ctrl-S for those, but we don't use them
that often so let's leave it where it is.