.bashrc become(): Switch back to the dir in which become was run
become() uses `sudo -i` in order to get a login shell, but this loses
our current working directory, which is annoying. We now pass in the
CWD as an environment variable and have full-env.bash switch to that
directory before finishing to continue on with the new shell.

Note that the environment variable must not be set before running sudo
(because sudo will usually remove it due to the standard config's
environment variable filtering) but instead be passed as a parameter
to sudo.